requirements.txt
asgiref==3.4.1
autopep8==1.6.0
dj-database-url==0.5.0
Django==4.0
django-cors-headers==3.10.1
djangorestframework==3.13.1
drf-writable-nested==0.6.3
gunicorn==20.1.0
mysqlclient==2.1.0
psycopg2==2.9.3
pycodestyle==2.8.0
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
